@@ -32,53 +32,57 @@
   </distributionManagement>
 
   <properties>
-    <jdkLevel>11</jdkLevel>
+    <jdkLevel>17</jdkLevel>
 
     <!-- GBIF libraries -->
-    <gbif-cli.version>1.1</gbif-cli.version>
     <gbif-api.version>2.0.0</gbif-api.version>
     <gbif-common-ws.version>2.0.3</gbif-common-ws.version>
     <gbif-registry.version>4.0.1</gbif-registry.version>
     <datacite-rest-client.version>2.0.0</datacite-rest-client.version>
 
+    <!-- Core libraries -->
     <lombok.version>1.18.32</lombok.version>
-    <slf4j.version>1.7.24</slf4j.version>
-    <guava.version>17.0</guava.version>
-    <jcommander.version>1.47</jcommander.version>
-    <rxjava1.version>1.1.2</rxjava1.version>
-    <validation-api.version>2.0.1.Final</validation-api.version>
-    <log4j.version>2.8</log4j.version>
-    <oltu.version>1.0.2</oltu.version>
-    <httpclient.version>4.5.3</httpclient.version>
-    <httpcore.version>4.4.12</httpcore.version>
-    <jackson.version>2.11.3</jackson.version>
-    <elasticsearch.version>7.10.0</elasticsearch.version>
-    <rxjava.version>2.0.0</rxjava.version>
-    <reactive-streams.version>1.0.0</reactive-streams.version>
-    <retrofit.version>2.3.0</retrofit.version>
-    <commons-lang3.version>3.4</commons-lang3.version>
-    <gson.version>2.7</gson.version>
-    <commons-io.version>2.4</commons-io.version>
-    <okio.version>1.9.0</okio.version>
+    <slf4j.version>1.7.36</slf4j.version>
+    <guava.version>33.2.1-jre</guava.version>
+    <jcommander.version>1.82</jcommander.version>
+    <validation-api.version>3.0.2</validation-api.version>
+    <log4j.version>2.20.0</log4j.version>
+    <logback.version>1.2.12</logback.version>
+    <httpclient.version>4.5.14</httpclient.version>
+    <httpcore.version>4.4.16</httpcore.version>
+    <jackson.version>2.15.2</jackson.version>
+    <elasticsearch.version>9.0.0</elasticsearch.version>
+    <rxjava.version>3.1.7</rxjava.version>
+    <reactive-streams.version>1.0.4</reactive-streams.version>
+    <retrofit.version>2.9.0</retrofit.version>
+    <commons-lang3.version>3.13.0</commons-lang3.version>
+    <gson.version>2.10.1</gson.version>
+    <commons-io.version>2.13.0</commons-io.version>
+    <okio.version>3.5.0</okio.version>
     <findbugs.version>3.0.1</findbugs.version>
-    <spring.version>5.2.15.RELEASE</spring.version>
-    <spring-boot.version>2.3.12.RELEASE</spring-boot.version>
+    <spring-boot.version>3.2.8</spring-boot.version>
+    <spring.version>6.1.11</spring.version>
+    <picocli.version>4.7.7</picocli.version>
+    <jakarta.annotation.version>2.1.1</jakarta.annotation.version>
+    <hibernate-validator.version>8.0.1.Final</hibernate-validator.version>
+    <jakarta.el.version>4.0.2</jakarta.el.version>
 
-    <apache.beam.version>2.15.0</apache.beam.version>
-    <postgresql.version>9.4.1208.jre7</postgresql.version>
-    <hikari.version>3.2.0</hikari.version>
-
-    <!-- Caching -->
+    <!-- Database and caching -->
+    <apache.beam.version>2.49.0</apache.beam.version>
+    <postgresql.version>42.6.0</postgresql.version>
+    <hikari.version>5.0.1</hikari.version>
     <cache2k.version>2.6.1.Final</cache2k.version>
 
-    <!-- Contentful SDK is used for crawling and building the index -->
+    <!-- Contentful SDK -->
     <contentful-sdk.version>10.5.18</contentful-sdk.version>
-    <!-- Contentful Management SDK is used for backup procedures only -->
     <cma-sdk.version>3.4.12</cma-sdk.version>
-    <okhttp.version>3.4.1</okhttp.version>
-    <metainf-services.version>1.7</metainf-services.version>
+    <okhttp.version>4.11.0</okhttp.version>
+    <metainf-services.version>1.9</metainf-services.version>
+    
+    <!-- Maven plugins -->
     <maven-shade-plugin.version>3.6.0</maven-shade-plugin.version>
     <maven-dependency-plugin.version>3.1.2</maven-dependency-plugin.version>
+    <maven-compiler-plugin.version>3.11.0</maven-compiler-plugin.version>
   </properties>
 
   <build>
