@@ -13,49 +13,50 @@
  */
 package org.gbif.content.crawl.contentful.backup;
 
-import org.gbif.cli.BaseCommand;
-import org.gbif.cli.Command;
 import org.gbif.content.crawl.conf.ContentCrawlConfiguration;
 
+import java.io.File;
 import java.io.IOException;
+import java.util.concurrent.Callable;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 
 import org.kohsuke.MetaInfServices;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import picocli.CommandLine.Command;
+import picocli.CommandLine.Option;
+
 /**
  * Command to trigger the backup of Contentful.
  */
-@MetaInfServices(Command.class)
-public class ContentfulBackupCommand extends BaseCommand {
+@Command(name = "contentful-backup", description = "Backup Contentful data")
+@MetaInfServices(Callable.class)
+public class ContentfulBackupCommand implements Callable<Integer> {
   private static final Logger LOG = LoggerFactory.getLogger(ContentfulBackupCommand.class);
-  private final ContentCrawlConfiguration configuration = new ContentCrawlConfiguration();
+
+  @Option(names = {"-c", "--config"}, description = "Configuration file path", required = true)
+  private String configFile;
 
   /**
-   * Default constructor, sets the command name.
-   */
-  public ContentfulBackupCommand() {
-    super("contentful-backup");
-  }
-
-  /**
-   * Configuration object.
+   * Executes the backup.
    */
   @Override
-  protected ContentCrawlConfiguration getConfigurationObject() {
-    return configuration;
-  }
-
-  /**
-   * Executes and asynchronous crawl.
-   */
-  @Override
-  protected void doRun() {
+  public Integer call() {
     try {
+      // Load configuration from file
+      ObjectMapper mapper = new ObjectMapper(new YAMLFactory());
+      ContentCrawlConfiguration configuration = mapper.readValue(new File(configFile), ContentCrawlConfiguration.class);
+      
+      LOG.info("Starting Contentful backup with config: {}", configFile);
       new ContentfulBackup(configuration).run();
+      LOG.info("Contentful backup completed successfully");
+      return 0;
     } catch (IOException e) {
       LOG.error("Backup failed!", e);
-      System.exit(1); // to enable triggering of e.g. an email from a Cron
+      return 1;
     }
   }
 }
