@@ -13,40 +13,48 @@
  */
 package org.gbif.content.crawl.contentful.crawl;
 
-import org.gbif.cli.BaseCommand;
-import org.gbif.cli.Command;
 import org.gbif.content.crawl.conf.ContentCrawlConfiguration;
 
+import java.io.File;
+import java.util.concurrent.Callable;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import org.kohsuke.MetaInfServices;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import picocli.CommandLine.Command;
+import picocli.CommandLine.Option;
 
 /**
- * Contentful crawl command. Triggers  crawl using the specified configuration.
+ * Contentful crawl command. Triggers crawl using the specified configuration.
  */
-@MetaInfServices(Command.class)
-public class ContentfulCrawlCommand extends BaseCommand {
+@Command(name = "contentful-crawl", description = "Crawl Contentful and index to Elasticsearch")
+@MetaInfServices(Callable.class)
+public class ContentfulCrawlCommand implements Callable<Integer> {
 
-  private  final ContentCrawlConfiguration configuration = new ContentCrawlConfiguration();
+  private static final Logger LOG = LoggerFactory.getLogger(ContentfulCrawlCommand.class);
 
-  /**
-   * Default constructor, sets the command name.
-   */
-  public ContentfulCrawlCommand() {
-    super("contentful-crawl");
-  }
+  @Option(names = {"-c", "--config"}, description = "Configuration file path", required = true)
+  private String configFile;
 
   /**
-   * Configuration object.
+   * Executes the crawl.
    */
   @Override
-  protected ContentCrawlConfiguration getConfigurationObject() {
-    return configuration;
-  }
-
-  /**
-   * Executes and asynchronous crawl.
-   */
-  @Override
-  protected void doRun() {
-    new ContentfulCrawler(configuration).run();
+  public Integer call() {
+    try {
+      // Load configuration from file
+      ObjectMapper mapper = new ObjectMapper(new YAMLFactory());
+      ContentCrawlConfiguration configuration = mapper.readValue(new File(configFile), ContentCrawlConfiguration.class);
+      
+      LOG.info("Starting Contentful crawl with config: {}", configFile);
+      new ContentfulCrawler(configuration).run();
+      LOG.info("Contentful crawl completed successfully");
+      return 0;
+    } catch (Exception e) {
+      LOG.error("Contentful crawl failed!", e);
+      return 1;
+    }
   }
 }
