@@ -14,13 +14,16 @@
 package org.gbif.content.crawl.contentful.crawl;
 
 import java.util.Collection;
-import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
 
-import org.elasticsearch.action.update.UpdateRequest;
-import org.elasticsearch.client.RequestOptions;
-import org.elasticsearch.client.RestHighLevelClient;
-import org.elasticsearch.script.Script;
-import org.elasticsearch.script.ScriptType;
+import co.elastic.clients.elasticsearch.ElasticsearchClient;
+import co.elastic.clients.elasticsearch.core.UpdateRequest;
+import co.elastic.clients.elasticsearch._types.Script;
+import co.elastic.clients.elasticsearch._types.ScriptSource;
+
+
+import co.elastic.clients.json.JsonData;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -43,9 +46,9 @@ public class ESDocumentLinker {
 
   private final String targetContentTypeId;
 
-  private final RestHighLevelClient esClient;
+  private final ElasticsearchClient esClient;
 
-  public ESDocumentLinker(String targetContentTypeId, RestHighLevelClient esClient) {
+  public ESDocumentLinker(String targetContentTypeId, ElasticsearchClient esClient) {
     this.esClient = esClient;
     this.targetContentTypeId =  targetContentTypeId;
   }
@@ -65,7 +68,7 @@ public class ESDocumentLinker {
    * Accepts list of localized resources and a single CDAEntry.
    */
   public void processEntryTag(Object entry, String esTypeName, String tagValue) {
-    if (Collection.class.isInstance(entry)) {
+    if (entry instanceof Collection) {
       processEntryTag((Collection<?>)entry, esTypeName, tagValue);
     } else {
       processEntryTag((CDAEntry)entry, esTypeName, tagValue);
@@ -77,7 +80,7 @@ public class ESDocumentLinker {
    */
   private void processEntryTag(Collection<?> resources, String esTypeName, String tagValue) {
     resources.stream()
-      .filter(resource -> CDAEntry.class.isInstance(resource)
+      .filter(resource -> resource instanceof CDAEntry
                           && ((CDAEntry) resource).contentType().id().equals(targetContentTypeId))
       .forEach(cdaEntry -> insertTag((CDAEntry) cdaEntry, esTypeName, tagValue));
   }
@@ -87,16 +90,33 @@ public class ESDocumentLinker {
    */
   private void insertTag(CDAEntry cdaEntry, String esTypeName, String tagValue) {
     try {
-      Script script = new Script(ScriptType.INLINE, Script.DEFAULT_SCRIPT_LANG,
-                                 String.format(NEWS_UPDATE_SCRIPT, esTypeName + "Tag"),
-                                 Collections.singletonMap("tag", tagValue));
-      UpdateRequest updateRequest = new UpdateRequest()
-                                      .index(getEsIdxName(cdaEntry.contentType().name()))
-                                      .id(cdaEntry.id())
-                                      .script(script);
-      esClient.update(updateRequest, RequestOptions.DEFAULT);
+      String indexName = getEsIdxName(cdaEntry.contentType().name());
+      String documentId = cdaEntry.id();
+      String fieldName = esTypeName + "Tag";
+      String scriptSource = String.format(NEWS_UPDATE_SCRIPT, fieldName);
+
+      Map<String, JsonData> params = new HashMap<>();
+      params.put("tag", JsonData.of(tagValue));
+
+      Script script = Script.of(s -> s
+          .lang("painless")
+          .source(ScriptSource.of( sc -> sc.scriptString(scriptSource)))
+          .params(params)
+      );
+
+      UpdateRequest<Object, Object> updateRequest = UpdateRequest.of(u -> u
+          .index(indexName)
+          .id(documentId)
+          .script(script)
+          .retryOnConflict(3)
+      );
+
+      esClient.update(updateRequest, Object.class);
+
+      LOG.info("Updated tag {} for entry {} in index {}", tagValue, documentId, indexName);
+
     } catch (Exception ex) {
-      LOG.error("Error updating news tag {} from entry {} ", tagValue, cdaEntry, ex);
+      LOG.error("Error updating news tag {} from entry {} ", tagValue, cdaEntry.id(), ex);
     }
   }
 }
