@@ -13,7 +13,6 @@
  */
 package org.gbif.content.crawl.contentful.crawl;
 
-import java.io.IOException;
 import java.util.Collection;
 import java.util.EnumSet;
 import java.util.HashMap;
@@ -24,14 +23,13 @@ import java.util.Set;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
-import org.elasticsearch.common.Strings;
-import org.elasticsearch.common.xcontent.XContentBuilder;
-import org.elasticsearch.common.xcontent.XContentFactory;
-
 import com.contentful.java.cma.Constants.CMAFieldType;
 import com.contentful.java.cma.model.CMAContentType;
 import com.contentful.java.cma.model.CMAField;
 import com.google.common.collect.ImmutableMap;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.node.ObjectNode;
+import com.fasterxml.jackson.databind.node.ArrayNode;
 
 /**
  * Generates the ElasticSearch mapping of a Contentful ContentType.
@@ -67,7 +65,7 @@ public class  MappingGenerator {
    * Fields that are common to all contentful content types.
    */
   private static final Map<String, String> KNOWN_FIELDS = new ImmutableMap.Builder<String, String>()
-    .put("_all",TEXT)
+    .put("search_text",TEXT)
     .put("locale", KEYWORD )
     .put("contentType", KEYWORD)
     .put("id", KEYWORD)
@@ -103,32 +101,23 @@ public class  MappingGenerator {
   private final Collection<String> vocabularies;
 
   /**
-   * Mapping properties.
-   * Creates a Json structure like:
-   * "properties": {
-   *   "fieldName": {
-   *     "type": type
-   *   }
-   * }
    * All KNOWN_FIELDS are mapped to data type 'keyword'.
    */
-  private static void addProperties(XContentBuilder mapping, Map<String, String> collapsedFields) throws IOException {
+  private static void addProperties(ObjectNode mapping, Map<String, String> collapsedFields) {
     Map<String,String> flatFields = new HashMap<>(collapsedFields);
     flatFields.putAll(KNOWN_FIELDS);
-    mapping.startObject("properties");
+    ObjectNode properties = mapping.putObject("properties");
     flatFields.forEach((key,type) -> {
         try {
-          mapping.startObject(key);
-            mapping.field("type", type);
+          ObjectNode fieldNode = properties.putObject(key);
+            fieldNode.put("type", type);
             if (KEYWORD.equals(type) || TEXT.equals(type)) {
-              mapping.field("copy_to", "_all");
+              fieldNode.put("copy_to", "search_text");
             }
-          mapping.endObject();
-        } catch (IOException ex) {
+        } catch (Exception ex) {
           throw new RuntimeException(ex);
         }
-      });
-    mapping.endObject();
+    });
   }
 
   /**
@@ -136,7 +125,7 @@ public class  MappingGenerator {
    * Produces the following JSON structure:
    * {
    *  "ignored_fields": {
-   *    "match": "images|space|revision|type",
+   *    "match": "space|revision|type",
    *    "match_pattern": "regex",
    *    "mapping": {
    *      "index": false
@@ -144,147 +133,114 @@ public class  MappingGenerator {
    *  }
    * }
    */
-  private static void addDefaultIgnoredFields(XContentBuilder mapping) throws IOException {
-    mapping.startObject();
-      mapping.startObject("ignored_fields");
-        mapping.field("match", IGNORED_FIELDS.pattern());
-        mapping.field("match_pattern", "regex");
-        mapping.startObject("mapping");
-          mapping.field("index",false);
-        mapping.endObject();
-      mapping.endObject();
-    mapping.endObject();
+  private static void addDefaultIgnoredFields(ArrayNode dynamicTemplate) {
+    // Simple ignored fields
+    ObjectNode template = dynamicTemplate.addObject();
+    ObjectNode ignoredFields = template.putObject("ignored_fields");
+    ignoredFields.put("match", IGNORED_FIELDS.pattern());
+    ignoredFields.put("match_pattern", "regex");
+    ObjectNode ignoredMapping = ignoredFields.putObject("mapping");
+    ignoredMapping.put("enabled", false);
   }
 
   /**
-   * Add mapping to reversed links.
-   * Produces the following JSON structure:
-   * {
-   *  "ignored_fields": {
-   *    "match": "*Tag",
-   *    "match_pattern": "regex",
-   *    "mapping": {
-   *      "type": "keyword"
-   *    }
-   *  }
-   * }
-   *
    * This is used to create links between content types, in particular in the News content type.
    */
-  private static void addGenericTagsMapping(XContentBuilder mapping) throws IOException {
-    mapping.startObject();
-      mapping.startObject("generic_tags");
-        mapping.field("match", ".*Tag");
-        mapping.field("match_pattern", "regex");
-        mapping.startObject("mapping");
-          mapping.field("type", KEYWORD);
-          mapping.field("copy_to", "_all");
-        mapping.endObject();
-      mapping.endObject();
-    mapping.endObject();
-  }
-
-
-  /**
-   * Assets mapping.
-   * Produces the following JSON structure:
-   * {
-   *  "asset_files": {
-   *    "path_match": "*.file.*",
-   *    "mapping": {
-   *      "index": "no"
-   *    }
-   *  }
-   * }
-   */
-  private static void addFileAssetMapping(XContentBuilder mapping) throws IOException {
-    mapping.startObject();
-      mapping.startObject("asset_files");
-        mapping.field("path_match", "*.file.*");
-        mapping.startObject("mapping");
-          mapping.field("index", false);
-        mapping.endObject();
-      mapping.endObject();
-    mapping.endObject();
+  private static void addGenericTagsMapping(ArrayNode dynamicTemplate) {
+    ObjectNode template = dynamicTemplate.addObject();
+    ObjectNode genericTags = template.putObject("generic_tags");
+      genericTags.put("match", ".*Tag");
+      genericTags.put("match_pattern", "regex");
+      ObjectNode genericTagsMapping = genericTags.putObject("mapping");
+        genericTagsMapping.put("type", KEYWORD);
+        genericTagsMapping.put("copy_to", "search_text");
   }
 
   /**
-   * Default nested mapping.
-   * Produces the following JSON structure:
+   * Adds the file asset mapping.
    * {
-   *  "asset_fieldName": {
-   *    "path_match": "*.fieldName.*",
-   *    "mapping": {
-   *      "type": esType
-   *    }
-   *  }
+   *   "asset_files": {
+   *     "path_match": "*.file.*",
+   *     "mapping": {
+   *       "index": false
+   *     }
+   *   }
    * }
    */
-  private static void addNestedMapping(XContentBuilder mapping, String fieldName, String esType) throws IOException {
-    mapping.startObject();
-      mapping.startObject("nested_" + fieldName);
-        mapping.field("path_match",  "*." + fieldName + ".*");
-        mapping.startObject("mapping");
-          mapping.field("type", esType);
-          if (KEYWORD.equals(esType) || TEXT.equals(esType)) {
-            mapping.field("copy_to", "_all");
-          }
-        mapping.endObject();
-      mapping.endObject();
-    mapping.endObject();
+  private static void addFileAssetMapping(ArrayNode dynamicTemplate) {
+    ObjectNode template = dynamicTemplate.addObject();
+    ObjectNode assetFiles = template.putObject("asset_files");
+      assetFiles.put("path_match", "*.file.*");
+      ObjectNode assetFilesMapping = assetFiles.putObject("mapping");
+        assetFilesMapping.put("type", "object");
+        assetFilesMapping.put("enabled", false);
   }
 
-    /**
-     * Default nested mapping.
-     * Produces the following JSON structure:
-     * {
-     *  "asset_fieldName": {
-     *    "path_match": "*.fieldName.*",
-     *    "mapping": {
-     *      "type": esType
-     *    }
-     *  }
-     * }
-     */
-    private static void addMatchMapping(XContentBuilder mapping, String fieldName, String esType) throws IOException {
-       mapping.startObject();
-          mapping.startObject("nested_" + fieldName);
-            mapping.field("match",  fieldName);
-            mapping.startObject("mapping");
-              mapping.field("type", esType);
-            mapping.endObject();
-          mapping.endObject();
-        mapping.endObject();
+  /**
+   * Adds nested mapping for a field.
+   * {
+   *   "nested_fieldName": {
+   *     "path_match": "*.fieldName.*",
+   *     "mapping": {
+   *       "type": "esType"
+   *     }
+   *   }
+   * }
+   */
+  private static void addNestedMapping(ArrayNode dynamicTemplate, String fieldName, String esType) {
+    ObjectNode template = dynamicTemplate.addObject();
+    ObjectNode nestedField = template.putObject("nested_" + fieldName);
+      nestedField.put("path_match",  "*." + fieldName + ".*");
+      ObjectNode nestedMapping = nestedField.putObject("mapping");
+        nestedMapping.put("type", esType);
+        if (KEYWORD.equals(esType) || TEXT.equals(esType)) {
+          nestedMapping.put("copy_to", "search_text");
+        }
+  }
+
+  /**
+   * Adds match mapping for a field.
+   * {
+   *   "nested_fieldName": {
+   *     "match": "fieldName",
+   *     "mapping": {
+   *       "type": "esType"
+   *     }
+   *   }
+   * }
+   */
+    private static void addMatchMapping(ArrayNode dynamicTemplate, String fieldName, String esType) {
+      ObjectNode template = dynamicTemplate.addObject();
+      ObjectNode nestedField = template.putObject("nested_" + fieldName);
+         nestedField.put("match",  fieldName);
+         ObjectNode nestedMapping = nestedField.putObject("mapping");
+           nestedMapping.put("type", esType);
     }
   /**
-   * Template for dynamic field mapping.
-   * Produces a Json object like:
+   * Adds a template field to the mapping.
    * {
    *   "fieldName": {
-   *     "match": fieldPattern,
+   *     "match": "fieldPattern",
    *     "mapping": {
-   *       "type": esType
+   *       "type": "esType"
    *     }
    *   }
    * }
    */
   public static void addTemplateField(String match, String fieldName, String fieldPattern, String esType,
-                                      XContentBuilder mapping) {
+                                       ArrayNode dynamicMapping) {
     try {
-      mapping.startObject();
-        mapping.startObject(fieldName);
-          mapping.field(match, fieldPattern);
-          mapping.startObject("mapping");
-            mapping.field("type", esType);
-            if (NESTED.equals(esType)) {
-                mapping.field("dynamic", true);
-            } else if (FULLTEXT_FIELDS.matcher(fieldName).matches()) {
-              mapping.field("copy_to", "_all");
-            }
-          mapping.endObject();
-        mapping.endObject();
-      mapping.endObject();
-    } catch (IOException ex) {
+      ObjectNode template = dynamicMapping.addObject();
+      ObjectNode fieldNode = template.putObject(fieldName);
+        fieldNode.put(match, fieldPattern);
+        ObjectNode fieldMapping = fieldNode.putObject("mapping");
+          fieldMapping.put("type", esType);
+          if (NESTED.equals(esType)) {
+              fieldMapping.put("dynamic", true);
+          } else if (FULLTEXT_FIELDS.matcher(fieldName).matches()) {
+            fieldMapping.put("copy_to", "search_text");
+          }
+    } catch (Exception ex) {
       throw new RuntimeException(ex);
     }
   }
@@ -348,42 +304,37 @@ public class  MappingGenerator {
    */
   public String getEsMapping(CMAContentType contentType) {
     Map<String, String> collapsedFields = new HashMap<>();
-    try (XContentBuilder mapping = XContentFactory.jsonBuilder()) {
-      mapping.startObject();
-        mapping.startArray("dynamic_templates");
-        addDefaultIgnoredFields(mapping);
-        addFileAssetMapping(mapping);
-        addGenericTagsMapping(mapping);
-        addNestedMapping(mapping, TITLE_FIELD, TEXT);
-        addNestedMapping(mapping, "description", TEXT);
-        addNestedMapping(mapping, "summary", TEXT);
-        addNestedMapping(mapping, "body", TEXT);
-        addNestedMapping(mapping, "title", TEXT);
-        addMatchMapping(mapping, "id", KEYWORD);
-        addMatchMapping(mapping, "isoCode", KEYWORD);
-        addNestedMapping(mapping, "label", KEYWORD);
-        addNestedMapping(mapping, "url", KEYWORD);
-        contentType.getFields().stream()
-          .filter(cmaField -> !cmaField.isDisabled()).forEach(cmaField ->
-          esType(cmaField).ifPresent(esType -> {
-            if (VOCABULARY.equals(esType)) {
-              collapsedFields.put(cmaField.getId(), KEYWORD);
-            } else if (isSimpleField(cmaField, esType)) {
-              collapsedFields.put(cmaField.getId(), esType);
-            } else if (!NESTED.equals(esType) && cmaField.isLocalized() && !cmaField.getName().equalsIgnoreCase("blocks")) { //localizable fields have nested elements
-              addTemplateField("path_match", cmaField.getId(), cmaField.getId() + ".*", esType, mapping);
-            } else if (!cmaField.getName().equalsIgnoreCase("blocks")) {
-              addTemplateField("match", cmaField.getId(), cmaField.getId(), esType, mapping);
-            }
-          })
-        );
-        mapping.endArray();
-        addProperties(mapping, collapsedFields);
-      mapping.endObject();
-      return Strings.toString(mapping);
-    } catch (IOException ex) {
-      throw new RuntimeException(ex);
-    }
+    ObjectMapper mapper = new ObjectMapper();
+    ObjectNode mapping = mapper.createObjectNode();
+    ArrayNode dynamicTemplates = mapping.putArray("dynamic_templates");
+    addDefaultIgnoredFields(dynamicTemplates);
+    addFileAssetMapping(dynamicTemplates);
+    addGenericTagsMapping(dynamicTemplates);
+    addNestedMapping(dynamicTemplates, TITLE_FIELD, TEXT);
+    addNestedMapping(dynamicTemplates, "description", TEXT);
+    addNestedMapping(dynamicTemplates, "summary", TEXT);
+    addNestedMapping(dynamicTemplates, "body", TEXT);
+    addNestedMapping(dynamicTemplates, "title", TEXT);
+    addMatchMapping(dynamicTemplates, "id", KEYWORD);
+    addMatchMapping(dynamicTemplates, "isoCode", KEYWORD);
+    addNestedMapping(dynamicTemplates, "label", KEYWORD);
+    addNestedMapping(dynamicTemplates, "url", KEYWORD);
+    contentType.getFields().stream()
+      .filter(cmaField -> !cmaField.isDisabled()).forEach(cmaField ->
+      esType(cmaField).ifPresent(esType -> {
+        if (VOCABULARY.equals(esType)) {
+          collapsedFields.put(cmaField.getId(), KEYWORD);
+        } else if (isSimpleField(cmaField, esType)) {
+          collapsedFields.put(cmaField.getId(), esType);
+        } else if (!NESTED.equals(esType) && cmaField.isLocalized() && !cmaField.getName().equalsIgnoreCase("blocks")) { //localizable fields have nested elements
+          addTemplateField("path_match", cmaField.getId(), cmaField.getId() + ".*", esType, dynamicTemplates);
+        } else if (!cmaField.getName().equalsIgnoreCase("blocks")) {
+          addTemplateField("match", cmaField.getId(), cmaField.getId(), esType, dynamicTemplates);
+        }
+      })
+    );
+    addProperties(mapping, collapsedFields);
+    return mapping.toPrettyString();
   }
 
   /**
