@@ -16,20 +16,17 @@ package org.gbif.content.crawl.mendeley.clients;
 import org.gbif.content.crawl.conf.ContentCrawlConfiguration;
 import org.gbif.content.crawl.es.ElasticSearchUtils;
 
-import java.io.Closeable;
-import java.io.IOException;
+import java.util.Arrays;
 import java.util.HashMap;
+import java.util.Map;
 import java.util.Optional;
 
 import org.cache2k.Cache;
 import org.cache2k.Cache2kBuilder;
-import org.elasticsearch.action.search.SearchRequest;
-import org.elasticsearch.action.search.SearchResponse;
-import org.elasticsearch.client.RequestOptions;
-import org.elasticsearch.client.RestHighLevelClient;
-import org.elasticsearch.index.query.QueryBuilders;
-import org.elasticsearch.search.SearchHit;
-import org.elasticsearch.search.builder.SearchSourceBuilder;
+import co.elastic.clients.elasticsearch.ElasticsearchClient;
+import co.elastic.clients.elasticsearch.core.SearchRequest;
+import co.elastic.clients.elasticsearch.core.SearchResponse;
+import co.elastic.clients.elasticsearch.core.search.Hit;
 
 import lombok.Builder;
 import lombok.Data;
@@ -39,7 +36,7 @@ import lombok.SneakyThrows;
 /**
  * Elasticsearch client to get information from project and programme indices.
  */
-public class ContentEsClient  implements Closeable {
+public class ContentEsClient {
 
   /**
    * Data obtained from Elasticsearch.
@@ -53,7 +50,7 @@ public class ContentEsClient  implements Closeable {
 
   }
 
-  private final RestHighLevelClient esClient;
+  private final ElasticsearchClient esClient;
 
   private final Cache<String, Optional<ProjectResponse>> cache;
 
@@ -68,15 +65,12 @@ public class ContentEsClient  implements Closeable {
 
 
   private static SearchRequest buildSearchByIdsRequest(String index, String[] fetchFields,String id) {
-    SearchSourceBuilder searchSourceBuilder = new SearchSourceBuilder()
-      .size(1) //only one result is expected
-      .fetchSource(fetchFields, null)
-      .query(QueryBuilders.idsQuery().addIds(id));
-
-    SearchRequest searchRequest = new SearchRequest();
-    searchRequest.indices(index);
-    searchRequest.source(searchSourceBuilder);
-    return searchRequest;
+    return new SearchRequest.Builder()
+        .size(1) //only one result is expected
+        .source(s -> s.filter(f -> f.includes(Arrays.asList(fetchFields))))
+        .query(q -> q.ids(ids -> ids.values(id)))
+        .index(index)
+        .build();
   }
 
   /**
@@ -87,10 +81,11 @@ public class ContentEsClient  implements Closeable {
     if (programmeId != null) {
       SearchRequest searchRequest = buildSearchByIdsRequest("programme", new String[]{"acronym"}, programmeId);
 
-      SearchResponse searchResponse = esClient.search(searchRequest, RequestOptions.DEFAULT);
+      SearchResponse<Object> searchResponse = esClient.search(searchRequest, Object.class);
 
-      if (searchResponse.getHits().getTotalHits().value > 0) { //are there results?
-        Object acronym = searchResponse.getHits().getAt(0).getSourceAsMap().get("acronym");
+      if (searchResponse.hits().total().value() > 0) { //are there results?
+        Map<String, Object> source = (Map<String, Object>) searchResponse.hits().hits().get(0).source();
+        Object acronym = source.get("acronym");
         if (acronym != null) {// hasAcronym
           return acronym.toString();
         }
@@ -102,9 +97,10 @@ public class ContentEsClient  implements Closeable {
   /**
    * Extracts the programme.id from the response.
    */
-  private static String getProgrammeId(SearchHit searchHit) {
-    if (searchHit.getSourceAsMap().containsKey("programme")) {
-      HashMap<String,?> project = (HashMap<String,?>)searchHit.getSourceAsMap().get("programme");
+  private static String getProgrammeId(Hit<Object> searchHit) {
+    Map<String, Object> source = (Map<String, Object>) searchHit.source();
+    if (source.containsKey("programme")) {
+      HashMap<String,?> project = (HashMap<String,?>)source.get("programme");
       Object identifier = project.get("id");
       if (identifier != null) {
         return (String)identifier;
@@ -116,9 +112,9 @@ public class ContentEsClient  implements Closeable {
   /**
    * Converts the hit result to a ProjectResponse.
    */
-  private ProjectResponse toProjectResponse(SearchHit searchHit) {
+  private ProjectResponse toProjectResponse(Hit<Object> searchHit) {
     return ProjectResponse.builder()
-            .identifier(searchHit.getId())
+            .identifier(searchHit.id())
             .programmeAcronym(getProgrammeAcronym(getProgrammeId(searchHit)))
             .build();
   }
@@ -137,16 +133,11 @@ public class ContentEsClient  implements Closeable {
   private Optional<ProjectResponse> getFromElastic(String projectId) {
     SearchRequest searchRequest = buildSearchByIdsRequest("project", new String[]{"programme.id"}, projectId);
 
-    SearchResponse searchResponse = esClient.search(searchRequest, RequestOptions.DEFAULT);
-    if (searchResponse.getHits().getTotalHits().value > 0) {
-      return Optional.of(toProjectResponse(searchResponse.getHits().getAt(0)));
+    SearchResponse<Object> searchResponse = esClient.search(searchRequest, Object.class);
+    if (searchResponse.hits().total().value() > 0) {
+      return Optional.of(toProjectResponse(searchResponse.hits().hits().get(0)));
     }
     return Optional.empty();
-  }
-
-  @Override
-  public void close() throws IOException {
-    esClient.close();
   }
 
 }
